requirements.txt
Django==1.5.2
Jinja2==2.6
Markdown==2.2.1
PyYAML==3.10
Pygments==1.6
Sphinx==1.1.3
astroid==1.0.0
ccuapi==0.5.0
coverage==3.6
distribute==0.6.31
dj-database-url==0.2.1
dj-static==0.0.5
django-coverage==1.2.4
django-extensions==1.0.3
django-filter==0.5.4
djangorestframework==2.2.0
docopt==0.6.1
docutils==0.10
gevent==0.13.8
greenlet==0.4.0
gunicorn==0.17.2
hiredis==0.1.1
logilab-common==0.60.0
mock==1.0.1
newrelic
psycopg2==2.4.6
pylint==1.0.0
readline==6.2.4.1
redis==2.7.2
requests==1.1.0
selenium==2.33.0
sh==1.08
splinter==0.5.2
static==0.4
statsd==1.0.0
suds==0.4
termcolor==1.1.0
wsgiref==0.1.2
-e git+http://github.com/dfuentes77/coveralls-python@d3ad40778dafa3683e2afa5eaa7dc0193fa64a19#egg=coveralls

